UI/BE: Refactor (change directory structure)
