+++ HoloRepositoryUI-Server/.cfignore
@@ -1,3 +1,3 @@
 node_modules
-/server
+/src
 /public
