+++ HoloRepositoryUI-Server/src/api/routes/imagingStudySeries/imagingStudySeries.router.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import controller from "../controllers/imagingStudySeries.controller";
+import controller from "./imagingStudySeries.controller";
 
 const ImagingStudySeriesRouter = Router()
   .get("/", controller.getAll)
