+++ HoloRepositoryUI-Server/src/api/routes/pipelines/pipelines.router.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import PipelinesController from "../controllers/pipelines.controller";
+import PipelinesController from "./pipelines.controller";
 
 const PipelinesRouter = Router().get("/", PipelinesController.getAll);
