+++ HoloRepositoryUI-Server/src/api/routes/practitioners/practitioners.service.ts
@@ -1,7 +1,7 @@
-import logger from "../../common/logger";
-import { IPractitioner } from "../../../../HoloRepositoryUI-Types";
+import logger from "../../../common/logger";
+import { IPractitioner } from "../../../../../HoloRepositoryUI-Types";
 
-import samplePractitioner from "../../../test/samples/samplePractitioner.json";
+import samplePractitioner from "../../../__tests__/samples/samplePractitioner.json";
 
 export class PractitionersService {
   public getById(pid: string): Promise<IPractitioner> {
