+++ HoloRepositoryUI-Server/src/routes.ts
@@ -1,9 +1,9 @@
 import { Application } from "express";
-import PatientsRouter from "./api/routers/patients.router";
-import HologramsRouter from "./api/routers/holograms.router";
-import PipelinesRouter from "./api/routers/pipelines.router";
-import ImagingStudySeriesRouter from "./api/routers/imagingStudySeries.router";
-import PractitionersRouter from "./api/routers/practitioners.router";
+import PatientsRouter from "./api/routes/patients/patients.router";
+import HologramsRouter from "./api/routes/holograms/holograms.router";
+import PipelinesRouter from "./api/routes/pipelines/pipelines.router";
+import ImagingStudySeriesRouter from "./api/routes/imagingStudySeries/imagingStudySeries.router";
+import PractitionersRouter from "./api/routes/practitioners/practitioners.router";
 
 const apiVersion = 1;
 const apiPrefix = `/api/v${apiVersion}`;
